refactor(engagement): move views + urls into dojo/engagement/ui/ [engagement Phase 5]
