+++ dojo/asset/urls.py
@@ -1,7 +1,7 @@
 from django.conf import settings
 from django.urls import re_path
 
-from dojo.engagement import views as dojo_engagement_views
+from dojo.engagement.ui import views as dojo_engagement_views
 from dojo.product import views
 from dojo.utils import redirect_view
